Add hipSYCL support Fix bad new/free pairs

+++ test.sh
@@ -7,13 +7,31 @@ export CUDA_DIR="$NVHPC_DIR/cuda/"
 export PATH=$NVHPC_DIR/compilers/bin/:${PATH:-}
 export KOKKOS_DIR="/home/tom/Downloads/kokkos-4.0.01/"
 
+export HIPSYCL_DEBUG_LEVEL=1
+
 export CPU_RANKS=16
 export GPU_RANKS=3
 
 VERBOSE="ON"
 PROBLEM="tea.in"
 
+declare -A enabled_models=(
+  ["serial"]=true
+  ["omp"]=true
+  ["omp-target"]=true
+  ["kokkos"]=true
+  ["cuda"]=true
+  ["hip"]=true
+  ["std-indices"]=true
+  ["sycl-acc"]=true
+  ["sycl-usm"]=true
+)
+
 function test() {
+  if [[ "${enabled_models[$3]}" != true ]]; then
+    echo "# Skipping model $3"
+    return
+  fi
   rm -rf "build_$3" # /CMakeCache.txt
   echo "$2" "${@:4}"
   cmake -DCMAKE_BUILD_TYPE=Release -S. -B "build_$3" -GNinja -DCMAKE_VERBOSE_MAKEFILE=$VERBOSE -DENABLE_MPI=ON -DMODEL="$3" "${@:4}" # &>/dev/null
@@ -34,14 +52,17 @@ function test() {
 }
 
 function test_nompi() {
+  if [[ "${enabled_models[$2]}" != true ]]; then
+    echo "# Skipping model $2"
+    return
+  fi
   rm -rf "build_$2" # /CMakeCache.txt
   echo "$2" "${@:3}"
   cmake -DCMAKE_BUILD_TYPE=Release -S. -B "build_$2" -GNinja -DCMAKE_VERBOSE_MAKEFILE=$VERBOSE -DENABLE_MPI=OFF -DMODEL="$2" "${@:3}" # &>/dev/null
   cmake --build "build_$2"                                                                                                            # &>/dev/null
-
   export OMP_TARGET_OFFLOAD=MANDATORY
-  export OMP_PLACES=cores
-  export OMP_PROC_BIND=true
+  #  export OMP_PLACES=cores
+  #  export OMP_PROC_BIND=true
   export OMP_NUM_THREADS=$(nproc)
   # | grep -i -e "This run" -e "Timestep *" #
   export ASAN_OPTIONS=detect_leaks=0
@@ -49,14 +70,23 @@ function test_nompi() {
   build_$2/*-tealeaf --device "$1" --file "$PROBLEM" --profile true --staging-buffer auto
 }
 
-#LIBOMPTARGET_INFO=-1
-# gdb  -batch -ex  run  -ex  bt  -ex  quit  --args
+#export CCACHE_DISABLE=1
+#
+##test_nompi "0" cuda -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CUDA_COMPILER=nvcc -DCUDA_ARCH=sm_60
+#export HIPSYCL_TARGETS="omp.accelerated"
+##export HIPSYCL_TARGETS="cuda:sm_60"
+#test_nompi "0" std-indices -DCMAKE_CXX_COMPILER="/opt/hipsycl/56f7219/bin/syclcc" -DCXX_EXTRA_FLAGS="--opensycl-stdpar;--opensycl-stdpar-unconditional-offload" -DCXX_EXTRA_LIBRARIES=tbb  # -DCMAKE_CXX_STANDARD=17 -DUSE_TBB=ON
+#
+#exit 0
+
+
 
 (
-
-  test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=g++
-  test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=clang++
-
+  (
+    :
+    test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=g++
+    test_nompi "0" serial -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=clang++
+  )
   (
     :
     test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=g++ -DCXX_EXTRA_FLAGS="-Ofast" -DUSE_TBB=ON
@@ -94,6 +124,55 @@ function test_nompi() {
     test_nompi "0" hip -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc
     test_nompi "0" hip -DMANAGED_ALLOC=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc # fails validation at > O1 lol
     test_nompi "0" hip -DSYNC_ALL_KERNELS=ON -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=hipcc
+  )
+  (
+    :
+    export HIPSYCL_TARGETS="omp.library-only"
+    test_nompi "OpenMP" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "OpenMP" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF  #?-DENABLE_PROFILING=ON
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
+    set -eu
+    export DPCPP_CPU_NUM_CUS=1
+    export DPCPP_CPU_SCHEDULE=static
+    test_nompi "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test_nompi "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+    test_nompi "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test_nompi "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
+    set -eu
+    cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
+    test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test_nompi "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+
+    export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
+    hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
+    test_nompi "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test_nompi "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"  #?-DENABLE_PROFILING=ON
+    test_nompi "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"  #?-DENABLE_PROFILING=ON
   )
   wait
 )
@@ -138,44 +217,59 @@ function test_nompi() {
     test $GPU_RANKS "0" hip -DMANAGED_ALLOC=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc # fails validation at > O1 lol
     test $GPU_RANKS "0" hip -DMANAGED_ALLOC=ON -DSYNC_ALL_KERNELS=ON -DCXX_EXTRA_FLAGS="-O1" -DCMAKE_CXX_COMPILER=hipcc
   )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
+    set -eu
+    export DPCPP_CPU_NUM_CUS=1
+    export DPCPP_CPU_SCHEDULE=static
+    test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+    test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=ON
+    test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
+  )
+  (
+    :
+    export HIPSYCL_TARGETS="omp.library-only"
+    test $CPU_RANKS "OpenMP" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $CPU_RANKS "OpenMP" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF  #?-DENABLE_PROFILING=ON
+  )
+  (
+    :
+    set +eu
+    module load mpi
+    source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
+    source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
+    set -eu
+    cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
+    test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
+
+    export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
+    hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
+    test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"  #?-DENABLE_PROFILING=ON
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"  #?-DENABLE_PROFILING=ON
+  )
   wait
 )
-#exit 0
-#
-(
-  set +eu
-  module load mpi
-  source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
-  source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh
-  set -eu
-  export DPCPP_CPU_NUM_CUS=1
-  export DPCPP_CPU_SCHEDULE=static
-  test $CPU_RANKS "Ryzen" sycl-acc -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
-  test $CPU_RANKS "Ryzen" sycl-usm -DSYCL_COMPILER=ONEAPI-ICPX -DUSE_HOSTTASK=OFF
-)
 
-(
-  set +eu
-  module load mpi
-  source /opt/intel/oneapi/tbb/2021.10.0/env/vars.sh
-  source /opt/intel/oneapi/compiler/2023.2.0/env/vars.sh --include-intel-llvm
-  set -eu
-
-  cuda_sycl_flags="-fsycl-targets=nvptx64-nvidia-cuda;-Xsycl-target-backend;--cuda-gpu-arch=sm_60;--cuda-path=$CUDA_DIR"
-  test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-  test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$cuda_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$cuda_sycl_flags"
-
-  export LD_LIBRARY_PATH=/opt/rocm-5.4.3/lib:${LD_LIBRARY_PATH:-}
-  hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
-  test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-  test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=OFF -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-)
+# exit 0
 
 ### CUDA-aware MPI ###
 (
@@ -193,7 +287,17 @@ function test_nompi() {
     #    test $GPU_RANKS "0" hip  -DCXX_EXTRA_FLAGS="-Ofast" -DCMAKE_CXX_COMPILER=/usr/lib/aomp_17.0-1/bin/hipcc # doesn't work with NVHPC'S OpenMPI, segfaults at runtime
 
   )
-
+  (
+    :
+    export HIPSYCL_TARGETS="cuda:sm_60"
+    test $GPU_RANKS "NVIDIA" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "NVIDIA" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    export HSA_OVERRIDE_GFX_VERSION=10.1.2 #  https://github.com/RadeonOpenCompute/ROCm/issues/1756#issuecomment-1160386571
+    export HIPSYCL_TARGETS="hip:gfx1012"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=HIPSYCL -DSYCL_COMPILER_DIR="/opt/hipsycl/e7ec6ed" -DUSE_HOSTTASK=OFF  #?-DENABLE_PROFILING=ON
+  )
+    exit 0
   (
     :
     set +eu
@@ -210,10 +314,8 @@ function test_nompi() {
     hip_sycl_flags="-fsycl;-fsycl-targets=amdgcn-amd-amdhsa;-Xsycl-target-backend;--offload-arch=gfx1012"
     test $GPU_RANKS "0" std-indices -DCMAKE_CXX_COMPILER=clang++ -DUSE_ONEDPL=DPCPP -DCXX_EXTRA_FLAGS="-fsycl;$hip_sycl_flags;-O1" # again, needs -O1
 
-    # doesn't work with NVHPC'S OpenMPI, segfaults at runtime
-    # test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON  -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
-
-    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-acc -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON  -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"
+    test $GPU_RANKS "Radeon" sycl-usm -DSYCL_COMPILER=ONEAPI-Clang -DUSE_HOSTTASK=ON -DSYCL_COMPILER_DIR=/opt/intel/oneapi/compiler/2023.1.0/linux/bin-llvm/ -DCXX_EXTRA_FLAGS="$hip_sycl_flags" -DCXX_EXTRA_LINK_FLAGS="$hip_sycl_flags"  #?-DENABLE_PROFILING=ON
   )
 
 )
